+++ server/DrawOperation/DrawEclipse.h
@@ -29,8 +29,8 @@ private:
     int thinkness_;
     int xa_;
     int ya_;
-    int a_;
-    int b_;
+    float a_;
+    float b_;
 };
 }  // DrawOperation
 }  // Server
