+++ server/DrawOperation/DrawTransCircle.h
@@ -34,7 +34,7 @@ private:
     int thinkness_;
     int xa_;
     int ya_;
-    int r_;
+    float r_;
     float alpha_;
 };
 }  // DrawOperation
